Ultrasound: Fix off-screen display window frame size overwrite
In certain scenarios, the frame size of a window which resides in
the off-screen display could have been overwritten by the frame
size of a window in the main display.

This fix uses a fresh Rect object for the queried frame size to
make sure it does not get overwritten.

Change-Id: I11a52fafb1e7561d042f856f6f750e7519a64db1