Telephony: Fix wrongly displayed data icon.
Currently in MSimNetworkController, while handling the data connection
state change for a particular sub, the data state is not getting updated.
This happens due to the late arrival of data state change event when
compared to the dds sub getting updated in the database. This leads to
the wrong display of data icon on SUB0 when there is no data call on it.
Fix is to maintain data state and data network type per subscription in
MSimNetworkController and handle the data state changed event even for
non-dds sub.

Change-Id: I926d6fd4187974e23b29eac03fcc740a96618ac7
CRs-Fixed: 807485