FP2-1575: DO NOT MERGE. Truncate and ellipsize app name if t
DO NOT MERGE. Truncate and ellipsize app name if too long for dialog.

If the given app name is too long for the permissions dialog, then
it can push the warning that the application will be able to record
the screen below the fold, letting the app basically set its own
dialog message in a way that a user would be difficult to detect as
fraudulant.

Bug: 23345192
Change-Id: If5881ca75d5c155ef5174351d245dbc3abdaa584