FPII-2377 : Elevation of privilege vulnerability in Zygote process CVE-2016-3911 A-30143607
An elevation of privilege in the Zygote process could enable a local malicious
application to execute arbitrary code within the context of a privileged
process. This issue is rated as High because it could be used to gain local access to elevated
capabilities, which are not normally accessible to a third-party application.

Change-Id: I867fe5adeba1b6725bc1939405866de9681c80bf