FPII-2378 : Elevation of privilege vulnerability in framework APIs CVE-2016-3912 A-30202481
The BroadcastQueue.sendPendingBroadcastsLocked API checks the PID field of ProcessRecord when a pending broadcast
is to be dispatched in the attached process. If the process that was to receive this broadcast doesn't attach,
and later another process gets started with same PID, the second process could execute BroadcastReceiver, even
though it can have a different UID.
The fix is designed to prevent the accidental launch of a broadcast receiver in an incorrect app instance.

Change-Id: I7ee40b4ae825471675ced8b6bd515a5994dd5078