FPII-2480: Elevation of privilege vulnerability in Framework APIs CVE-2016-6715 A-29833954
An app can inject media keys, which allows an application to pick up 
the phone while it is ringing. This enables an attacker to call the device, 
have the malicious application pick up the phone, and thus enable
the attacker to listen to the users conversation without their consent.

Change-Id: I82de50a85367ab4f91c770a6a3b309c200bb0b1a