FPII-2470 : [Part 2/4] Update: Elevation of privilege vulnerability in Zygote process CVE-2016-3911 A-30963384
High
An elevation of privilege in the Zygote process could enable a local malicious application to execute arbitrary code within the context of a privileged process. This issue is rated as High because it could be used to gain local access to elevated capabilities, which are not normally accessible to a third-party application.
Note: This is an update to CVE-2016-3911, which was released as part of the September 2016 Partner Security Bulletin.
Additional technical details:
A-30963384	
The original fix targeted resource file descriptors. After further analysis, this issue affects all file descriptors leaked from Zygote to its children. A malicious process can modify the seek position of any of these file descriptors, and the seek location in all other processes could be affected.
The fix is designed to close the file descriptors and reopen them when the Zygote forks.

Change-Id: I034f2e26a55aa6435d0c080d04d5e4bda15093a5